update api route reset password
