+++ .env.development
@@ -1,4 +1,5 @@
 NEXT_PUBLIC_GRAPHQL_URI="http://localhost:3000/api/graphql"
+
 # Set to 1 when relying on a cross-domain API (provided by Vulcan Express, Vulcan Meteor backend or another Vulcan Next app for instance)
 # NEXT_PUBLIC_CROSS_DOMAIN_GRAPHQL_URI=1
