+++ src/lib/routes.ts
@@ -1,0 +1,10 @@
+export const routes = {
+  account: {
+    forgottenPassword: {
+      href: "/account/forgotten-password",
+    },
+    resetPassword: {
+      href: "/account/reset-password",
+    },
+  },
+};
